backend/requirements.txt
djangorestframework-3.10.3
django-cors-headers-3.2.0
django-filter-2.2.0
Pillow
#########################################
npm install ngx-pagination
npm install material-design-icons
ng add @angular/material
npm install --save ng2-file-upload

